pyproject.toml
[project]
name = "paperops"
version = "0.2.0"
description = "Utilities for academic papers, presentations, and publications"
readme = "README.md"
requires-python = ">=3.13"
dependencies = []

[project.optional-dependencies]
plotting = [
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "numpy>=1.24.0",
]
slides = [
    "python-pptx>=1.0.0",
    "cairosvg>=2.7.0",
    "Pillow>=10.0.0",
    "lxml>=4.9.0",
]
dev = ["pytest>=7.0.0", "black>=23.0.0", "isort>=5.12.0", "flake8>=6.0.0"]
all = [
    "paperops[plotting]",
    "paperops[slides]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py313']

[tool.isort]
profile = "black"
line_length = 88

[tool.hatch.build.targets.wheel]
packages = ["src/paperops"]
